+++ src/Api/Providers/Keywords.php
@@ -15,15 +15,14 @@ class Keywords extends Provider
         );
         $response = $this->getRequest()->exec(UrlHelper::getSearchUrl() . '?' . $get);
 
-        return $this->parseKeywordsFromRequest($response, $query);
+        return $this->parseKeywordsFromRequest($response);
     }
 
     /**
-     * @param string $query
      * @param array $response
      * @return array|null
      */
-    protected function parseKeywordsFromRequest($response, $query)
+    protected function parseKeywordsFromRequest($response)
     {
         if (!isset($response['resource_data_cache'][0]['data']['guides'])) {
             return null;
@@ -32,8 +31,12 @@ class Keywords extends Provider
         $keywords = $response['resource_data_cache'][0]['data']['guides'];
 
         return array_map(
-            function ($keywordData) use ($query) {
-                return $query . ' ' . $keywordData['term'];
+            function ($keywordData) {
+                return [
+                    'term'     => $keywordData['term'],
+                    'position' => $keywordData['position'],
+                    'display'  => $keywordData['display']
+                ];
             }, $keywords
         );
     }